chore: Move policies endpoint to the endpoint section of app
